backend/mcp_server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "mcp[cli]>=1.12.2",
    "pandas>=2.3.1",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
]
